test/shamhub: Split into multiple files
Everything in two files was starting to get quite ungainly.

[skip changelog]: Not user facing.
